terraform/challenge-02-aws-only/requirements.txt
PyJWT==2.8.0
requests==2.31.0
charset_normalizer==3.3.2
urllib3==2.0.7
certifi==2023.7.22
idna==3.4
